+++ src/Ext/TechnoType/Body.cpp
@@ -872,6 +872,64 @@ void TechnoTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	this->AutoTargetOwnPosition_Self.Read(exINI, pSection, "AutoFire.TargetSelf"); // Temporary solution for the INI tags renaming issue, see #2093
 	this->AutoTargetOwnPosition_Self.Read(exINI, pSection, "AutoTargetOwnPosition.Self");
 
+	this->AttachmentTopLayerMinHeight.Read(exINI, pSection, "AttachmentTopLayerMinHeight");
+	this->AttachmentUndergroundLayerMaxHeight.Read(exINI, pSection, "AttachmentUndergroundLayerMaxHeight");
+
+	// The following loop iterates over size + 1 INI entries so that the
+	// vector contents can be properly overriden via scenario rules - Kerbiter
+	for (size_t i = 0; i <= this->AttachmentData.size(); ++i)
+	{
+		char tempBuffer[32];
+		NullableIdx<AttachmentTypeClass> type;
+		_snprintf_s(tempBuffer, sizeof(tempBuffer), "Attachment%d.Type", i);
+		type.Read(exINI, pSection, tempBuffer);
+
+		if (!type.isset())
+			continue;
+
+		NullableIdx<TechnoTypeClass> technoType;
+		_snprintf_s(tempBuffer, sizeof(tempBuffer), "Attachment%d.TechnoType", i);
+		technoType.Read(exINI, pSection, tempBuffer);
+
+		Valueable<CoordStruct> flh;
+		_snprintf_s(tempBuffer, sizeof(tempBuffer), "Attachment%d.FLH", i);
+		flh.Read(exINI, pSection, tempBuffer);
+
+		Valueable<bool> isOnTurret;
+		_snprintf_s(tempBuffer, sizeof(tempBuffer), "Attachment%d.IsOnTurret", i);
+		isOnTurret.Read(exINI, pSection, tempBuffer);
+
+		Valueable<DirType> rotationAdjust;
+		_snprintf_s(tempBuffer, sizeof(tempBuffer), "Attachment%d.RotationAdjust", i);
+		rotationAdjust.Read(exINI, pSection, tempBuffer);
+
+		PhobosFixedString<32> id;
+		_snprintf_s(tempBuffer, sizeof(tempBuffer), "Attachment%d.ID", i);
+		id.Read(pINI, pSection, tempBuffer);
+
+		AttachmentDataEntry const entry { ValueableIdx<AttachmentTypeClass>(type), technoType, flh, isOnTurret, rotationAdjust, id };
+		if (i == AttachmentData.size())
+			this->AttachmentData.push_back(entry);
+		else
+			this->AttachmentData[i] = entry;
+	}
+
+	// Validate attachment ID uniqueness
+	std::set<PhobosFixedString<32>> usedIds;
+	for (size_t i = 0; i < this->AttachmentData.size(); ++i)
+	{
+		const auto& id = this->AttachmentData[i].ID;
+
+		if (!id)
+			continue;
+
+		if (!usedIds.insert(id).second)
+		{
+			Debug::FatalErrorAndExit("[%s] Duplicate Attachment ID '%s'\n",
+				pSection, id);
+		}
+	}
+
 	this->NoSecondaryWeaponFallback.Read(exINI, pSection, "NoSecondaryWeaponFallback");
 	this->NoSecondaryWeaponFallback_AllowAA.Read(exINI, pSection, "NoSecondaryWeaponFallback.AllowAA");
 	this->AllowWeaponSelectAgainstWalls.Read(exINI, pSection, "AllowWeaponSelectAgainstWalls");
@@ -1901,6 +1959,10 @@ void TechnoTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->Unsellable)
 
 		.Process(this->TurretShape)
+
+		.Process(this->AttachmentTopLayerMinHeight)
+		.Process(this->AttachmentUndergroundLayerMaxHeight)
+		.Process(this->AttachmentData)
 		;
 }
 void TechnoTypeExt::ExtData::LoadFromStream(PhobosStreamReader& Stm)
@@ -1914,6 +1976,33 @@ void TechnoTypeExt::ExtData::SaveToStream(PhobosStreamWriter& Stm)
 	Extension<TechnoTypeClass>::SaveToStream(Stm);
 	this->Serialize(Stm);
 }
+
+#pragma region Data entry save/load
+
+bool TechnoTypeExt::ExtData::AttachmentDataEntry::Load(PhobosStreamReader& stm, bool registerForChange)
+{
+	return this->Serialize(stm);
+}
+
+bool TechnoTypeExt::ExtData::AttachmentDataEntry::Save(PhobosStreamWriter& stm) const
+{
+	return const_cast<AttachmentDataEntry*>(this)->Serialize(stm);
+}
+
+template <typename T>
+bool TechnoTypeExt::ExtData::AttachmentDataEntry::Serialize(T& stm)
+{
+	return stm
+		.Process(this->Type)
+		.Process(this->TechnoType)
+		.Process(this->FLH)
+		.Process(this->IsOnTurret)
+		.Process(this->RotationAdjust)
+		.Process(this->ID)
+		.Success();
+}
+
+#pragma endregion
 
 // =============================
 // container