+++ src/Phobos.Ext.cpp
@@ -242,7 +242,10 @@ using PhobosTypeRegistry = TypeRegistry <
 	AttachEffectTypeClass,
 	AttachEffectClass,
 	NewSWType,
-	SelectBoxTypeClass
+	SelectBoxTypeClass,
+	AttachEffectClass,
+	AttachmentClass,
+	AttachmentTypeClass
 	// other classes
 > ;